refactor: move all git calls in utils into their own folder
