+++ src/actions/clean_branches.ts
@@ -9,8 +9,8 @@ import {
 } from '../lib/config/merge_conflict_callstack_config';
 import { TContext } from '../lib/context';
 import { KilledError } from '../lib/errors';
-import { checkoutBranch } from '../lib/utils/checkout_branch';
-import { getMergeBase } from '../lib/utils/merge_base';
+import { checkoutBranch } from '../lib/git/checkout_branch';
+import { getMergeBase } from '../lib/git/merge_base';
 import { logInfo, logTip } from '../lib/utils/splog';
 import { getTrunk } from '../lib/utils/trunk';
 import { Branch } from '../wrapper-classes/branch';
