+++ src/actions/edit_branch.ts
@@ -4,9 +4,9 @@ import {
   PreconditionsFailedError,
   RebaseConflictError,
 } from '../lib/errors';
+import { rebaseInProgress } from '../lib/git/rebase_in_progress';
 import { currentBranchPrecondition } from '../lib/preconditions';
 import { gpExecSync } from '../lib/utils/exec_sync';
-import { rebaseInProgress } from '../lib/utils/rebase_in_progress';
 import { rebaseUpstack } from './fix';
 
 export async function editBranchAction(context: TContext): Promise<void> {
