+++ src/actions/init.ts
@@ -3,8 +3,8 @@ import fs from 'fs-extra';
 import prompts from 'prompts';
 import { TContext } from '../lib/context';
 import { PreconditionsFailedError } from '../lib/errors';
+import { branchExists } from '../lib/git/branch_exists';
 import { currentGitRepoPrecondition } from '../lib/preconditions';
-import { branchExists } from '../lib/utils/branch_exists';
 import { logError, logInfo, logNewline } from '../lib/utils/splog';
 import { inferTrunk } from '../lib/utils/trunk';
 import { Branch } from '../wrapper-classes/branch';
