+++ src/actions/sync/merge_downstack.ts
@@ -2,9 +2,9 @@ import chalk from 'chalk';
 import prompts from 'prompts';
 import { TContext } from '../../lib/context';
 import { KilledError } from '../../lib/errors';
+import { copyFromRemote } from '../../lib/git/copy_from_remote';
+import { getMergeBase } from '../../lib/git/merge_base';
 import { syncPRInfoForBranchByName } from '../../lib/sync/pr_info';
-import { copyFromRemote } from '../../lib/utils/copy_from_remote';
-import { getMergeBase } from '../../lib/utils/merge_base';
 import { logInfo, logNewline, logWarn } from '../../lib/utils/splog';
 import { getTrunk } from '../../lib/utils/trunk';
 import { Branch } from '../../wrapper-classes/branch';
