+++ src/commands/stack-commands/test.ts
@@ -5,9 +5,9 @@ import yargs from 'yargs';
 import { validate } from '../../actions/validate';
 import { TContext } from '../../lib/context';
 import { ValidationFailedError } from '../../lib/errors';
+import { checkoutBranch } from '../../lib/git/checkout_branch';
 import { currentBranchPrecondition } from '../../lib/preconditions';
 import { profile } from '../../lib/telemetry/profile';
-import { checkoutBranch } from '../../lib/utils/checkout_branch';
 import { gpExecSync } from '../../lib/utils/exec_sync';
 import { logInfo } from '../../lib/utils/splog';
 import { getTrunk } from '../../lib/utils/trunk';
