+++ src/lib/git/add_all.ts
@@ -1,5 +1,5 @@
 import { ExitFailedError } from '../errors';
-import { gpExecSync } from './exec_sync';
+import { gpExecSync } from '../utils/exec_sync';
 
 export function addAll(): void {
   gpExecSync(
