+++ src/lib/git/branch_exists.ts
