+++ src/lib/git/checkout_branch.ts
@@ -1,5 +1,5 @@
 import { ExitFailedError } from '../errors';
-import { gpExecSync } from './exec_sync';
+import { gpExecSync } from '../utils/exec_sync';
 
 export function checkoutBranch(
   branch: string,
