+++ src/lib/git/commit.ts
@@ -1,6 +1,6 @@
 import { execStateConfig } from '../config/exec_state_config';
 import { ExitFailedError } from '../errors';
-import { gpExecSync } from './exec_sync';
+import { gpExecSync } from '../utils/exec_sync';
 
 const EMPTY_COMMIT_MESSAGE_INFO = [
   '\n',
