+++ src/lib/git/detect_staged_changes.ts
