+++ src/lib/git/git_status_utils.ts
@@ -1,5 +1,5 @@
 import { ExitFailedError } from '../errors';
-import { gpExecSync } from './exec_sync';
+import { gpExecSync } from '../utils/exec_sync';
 
 function doChangesExist(cmd: string): boolean {
   return (
