+++ src/lib/git/is_empty_branch.ts
