+++ src/lib/git/merge_base.ts
@@ -1,4 +1,4 @@
-import { gpExecSync } from './exec_sync';
+import { gpExecSync } from '../utils/exec_sync';
 
 export function getMergeBase(left: string, right: string): string {
   return gpExecSync({ command: `git merge-base ${left} ${right}` })
