+++ src/lib/git/rebase_in_progress.ts
